autotests/requirements.txt
pytest
pika
psycopg2-binary